+++ Scheduler+Agent/restart_scheduler.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root/sudo"
+  exit
+fi
+
 ps -ef | grep scheduler.py | grep -v grep | awk '{print $2}' | sudo xargs kill
 sudo kill `sudo lsof -t -i:52222` &>/dev/null
 echo "Restarting scheduler in some time.."
